Disable macOS jobs on TravisCI
Builds on macOS have been disabled because it's not possible to use
"brew link" with the icu4c formula anymore which makes compilation
fail for some reason (Crystal macro run doesnt seems to take in
account compilation env. var. correctly).

The library is not yet compatible with ICU version shipped with macOS.